Refactored RentedCar package Added JUnit tests

+++ src/main/java/com/epam/internship/carrental/rentedcar/RentedCarServiceImpl.java
@@ -1,5 +1,6 @@
 package com.epam.internship.carrental.rentedcar;
 
+import com.epam.internship.carrental.exceptions.RentedCarNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.domain.Page;
@@ -15,8 +16,6 @@ import java.util.Optional;
 @Qualifier("rentedCarService")
 public class RentedCarServiceImpl implements RentedCarService {
 
-    private static final String AUTH_HEADER = "Token ";
-    private static final String AUTH_TOKEN = AUTH_HEADER + "employee_token";
 
     private final RentedCarRepository rentedCarRepository;
 
@@ -29,31 +28,19 @@ public class RentedCarServiceImpl implements RentedCarService {
      * {@inheritDoc}
      */
     @Override
-    public void bookCarRentalWithAuthorization(final RentedCar rentedCar,
-                                                         final String authorization) {
-//        if (!authorization.equals(AUTH_TOKEN)) {
-//            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-//        }
-        rentedCarRepository.save(rentedCar);
-//        return new ResponseEntity<>(HttpStatus.OK);
+    public void bookCarRental(final RentedCarViewObject rentedCarViewObject) {
+        rentedCarRepository.save(RentedCarConverter.rentedCarFromRentedCarViewObject(rentedCarViewObject));
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public void endCarRentalWithAuthorization(final Long id,
-                                                        final String authorization) {
-//        if (!authorization.equals(AUTH_TOKEN)) {
-//            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-//        }
-//        Optional<RentedCar> optionalRentedCar = rentedCarRepository.findById(id);
-//        if (!optionalRentedCar.isPresent()) {
-//            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-//        } else {
-//
-//            return new ResponseEntity<>(HttpStatus.OK);
-//        }
+    public void endCarRental(final Long id) {
+        Optional<RentedCar> optionalRentedCar = rentedCarRepository.findById(id);
+        if (!optionalRentedCar.isPresent()) {
+            throw new RentedCarNotFoundException();
+        }
         rentedCarRepository.deleteById(id);
     }
 
@@ -61,20 +48,16 @@ public class RentedCarServiceImpl implements RentedCarService {
      * {@inheritDoc}
      */
     @Override
-    public void modifyCarRentalWithAuthorization(final Long id,
-                                                           final RentedCar rentedCar,
-                                                           final String authorization) {
-//        if (!authorization.equals(AUTH_TOKEN)) {
-//            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-//        }
+    public RentedCarViewObject modifyCarRental(final Long id,
+                                               final RentedCarViewObject rentedCarViewObject) {
         Optional<RentedCar> optionalRentedCar = rentedCarRepository.findById(id);
         if (!optionalRentedCar.isPresent()) {
-//            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-        }else {
+            throw new RentedCarNotFoundException();
+        } else {
             RentedCar modifiableRentedCar = optionalRentedCar.get();
-            RentedCar modifiedRentedCar = RentedCarUtilities.modifyRentedCar(modifiableRentedCar,rentedCar);
-            rentedCarRepository.save(modifiedRentedCar);
-//            return new ResponseEntity<>(rentedCarRepository.save(modifiedRentedCar),HttpStatus.OK);
+            RentedCar updatedRentedCar = RentedCarConverter.rentedCarFromRentedCarViewObject(rentedCarViewObject);
+            RentedCar modifiedRentedCar = RentedCarUtilities.modifyRentedCar(modifiableRentedCar, updatedRentedCar);
+            return RentedCarConverter.rentedCarViewObjectFromRentedCar(rentedCarRepository.save(modifiedRentedCar));
         }
     }
 
@@ -82,11 +65,8 @@ public class RentedCarServiceImpl implements RentedCarService {
      * {@inheritDoc}
      */
     @Override
-    public Page<RentedCar> listAllCarRentalWithAuthorization(final Pageable pageable,
-                                                                             final String authorization) {
-//        if (!authorization.equals(AUTH_TOKEN)) {
-//            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-//        }
-        return rentedCarRepository.findAll(pageable);
+    public Page<RentedCarViewObject> listAllCarRental(final Pageable pageable) {
+        return rentedCarRepository.findAll(pageable)
+                .map(RentedCarConverter::rentedCarViewObjectFromRentedCar);
     }
 }
